prune setup
