+++ lightning_transformers/__init__.py
@@ -7,7 +7,7 @@ __version__ = "0.2.4"
 __author__ = "Lightning AI et al."
 __author_email__ = "pytorch@lightning.ai"
 __license__ = "Apache-2.0"
-__copyright__ = f"Copyright (c) 2020-2020, {__author__}."
+__copyright__ = f"Copyright (c) 2020-2022, {__author__}."
 __homepage__ = "https://github.com/Lightning-AI/lightning-transformers"
 __docs__ = "Lightning Transformers."
 __long_doc__ = """
