+++ setup.py
@@ -1,22 +1,80 @@
 #!/usr/bin/env python
 import os
+import re
+from typing import List
 
 from setuptools import find_packages, setup
 
 import lightning_transformers as ltf
-from lightning_transformers import setup_tools
 
 _PATH_ROOT = os.path.dirname(__file__)
 _PATH_REQUIRE = os.path.join(_PATH_ROOT, "requirements")
 
-long_description = setup_tools._load_readme_description(_PATH_ROOT, homepage=ltf.__homepage__, ver=ltf.__version__)
+
+def _load_requirements(path_dir: str, file_name: str = "requirements.txt", comment_char: str = "#") -> List[str]:
+    """Load requirements from a file.
+
+    >>> _load_requirements(_PATH_ROOT)  # doctest: +ELLIPSIS +NORMALIZE_WHITESPACE
+    ['pytorch-lightning...', 'torch...', ...]
+    """
+    with open(os.path.join(path_dir, file_name)) as file:
+        lines = [ln.strip() for ln in file.readlines()]
+    reqs = []
+    for ln in lines:
+        # filer all comments
+        if comment_char in ln:
+            ln = ln[: ln.index(comment_char)].strip()
+        # skip directly installed dependencies
+        if ln.startswith("http"):
+            continue
+        if ln:  # if requirement is not empty
+            reqs.append(ln)
+    return reqs
+
+
+def _load_readme_description(path_dir: str, homepage: str, ver: str) -> str:
+    """Load readme as decribtion.
+
+    >>> _load_readme_description(_PATH_ROOT, "", "")  # doctest: +ELLIPSIS +NORMALIZE_WHITESPACE
+    '# ...'
+    """
+    path_readme = os.path.join(path_dir, "README.md")
+    text = open(path_readme, encoding="utf-8").read()
+
+    # https://github.com/PyTorchLightning/pytorch-lightning/raw/master/docs/source/_images/lightning_module/pt_to_pl.png
+    github_source_url = os.path.join(homepage, "raw", ver)
+    # replace relative repository path to absolute link to the release
+    #  do not replace all "docs" as in the readme we reger some other sources with particular path to docs
+    text = text.replace("docs/source/_static/", f"{os.path.join(github_source_url, 'docs/source/_static/')}")
+
+    # readthedocs badge
+    text = text.replace("badge/?version=stable", f"badge/?version={ver}")
+    text = text.replace(
+        "lightning-transformers.readthedocs.io/en/stable/", f"lightning-transformers.readthedocs.io/en/{ver}"
+    )
+    # codecov badge
+    text = text.replace("/branch/master/graph/badge.svg", f"/release/{ver}/graph/badge.svg")
+    # replace github badges for release ones
+    text = text.replace("badge.svg?branch=master&event=push", f"badge.svg?tag={ver}")
+
+    skip_begin = r"<!-- following section will be skipped from PyPI description -->"
+    skip_end = r"<!-- end skipping PyPI description -->"
+    # todo: wrap content as commented description
+    text = re.sub(rf"{skip_begin}.+?{skip_end}", "<!--  -->", text, flags=re.IGNORECASE + re.DOTALL)
+
+    # # https://github.com/Borda/pytorch-lightning/releases/download/1.1.0a6/codecov_badge.png
+    # github_release_url = os.path.join(homepage, "releases", "download", ver)
+    # # download badge and replace url with local file
+    # text = _parse_for_badge(text, github_release_url)
+    return text
+
 
 # https://setuptools.readthedocs.io/en/latest/setuptools.html#declaring-extras
 # Define package extras. These are only installed if you specify them.
 # From remote, use like `pip install lightning-transformers[extra]`
 # From local copy of repo, use like `pip install ".[extra]"`
 extras = {
-    "extra": setup_tools._load_requirements(path_dir=_PATH_REQUIRE, file_name="extra.txt"),
+    "extra": _load_requirements(path_dir=_PATH_REQUIRE, file_name="extra.txt"),
 }
 
 setup(
@@ -29,7 +87,7 @@ setup(
     download_url="https://github.com/Lightning-AI/lightning-transformers",
     license=ltf.__license__,
     packages=find_packages(exclude=["tests", "docs"]),
-    long_description=long_description,
+    long_description=_load_readme_description(_PATH_ROOT, homepage=ltf.__homepage__, ver=ltf.__version__),
     long_description_content_type="text/markdown",
     include_package_data=True,
     zip_safe=False,
@@ -37,7 +95,7 @@ setup(
     python_requires=">=3.7",
     setup_requires=[],
     extras_require=extras,
-    install_requires=setup_tools._load_requirements(_PATH_ROOT),
+    install_requires=_load_requirements(_PATH_ROOT),
     project_urls={
         "Bug Tracker": "https://github.com/Lightning-AI/lightning-transformers/issues",
         "Documentation": "https://lightning-transformers.readthedocs.io/en/stable/",
